Move case list refresh timer to other class
